+++ src/amr_criteria/amr_criteria.cpp
@@ -95,6 +95,26 @@ AmrTag AMRFirstDerivative::operator()(const MeshBlockData<Real> *rc) const {
   return Refinement::FirstDerivative(bnds, q, refine_criteria, derefine_criteria);
 }
 
+void AMRFirstDerivative::operator()(MeshData<Real> *md,
+                                    ParArray1D<AmrTag> &amr_tags) const {
+  auto ib = md->GetBoundsI(IndexDomain::interior);
+  auto jb = md->GetBoundsJ(IndexDomain::interior);
+  auto kb = md->GetBoundsK(IndexDomain::interior);
+  auto bnds = AMRBounds(ib, jb, kb);
+  auto dims = md->GetMeshPointer()->resolved_packages->FieldMetadata(field).Shape();
+  int n5(0), n4(0);
+  if (dims.size() > 2) {
+    n5 = dims[1];
+    n4 = dims[2];
+  } else if (dims.size() > 1) {
+    n5 = dims[0];
+    n4 = dims[1];
+  }
+  const int idx = comp4 + n4 * (comp5 + n5 * comp6);
+  Refinement::FirstDerivative(bnds, md, field, idx, amr_tags, refine_criteria,
+                              derefine_criteria, max_level);
+}
+
 AmrTag AMRSecondDerivative::operator()(const MeshBlockData<Real> *rc) const {
   if (!rc->HasVariable(field) || !rc->IsAllocated(field)) {
     return AmrTag::same;
@@ -103,6 +123,26 @@ AmrTag AMRSecondDerivative::operator()(const MeshBlockData<Real> *rc) const {
   auto q = Kokkos::subview(rc->Get(field).data, comp6, comp5, comp4, Kokkos::ALL(),
                            Kokkos::ALL(), Kokkos::ALL());
   return Refinement::SecondDerivative(bnds, q, refine_criteria, derefine_criteria);
+}
+
+void AMRSecondDerivative::operator()(MeshData<Real> *md,
+                                     ParArray1D<AmrTag> &amr_tags) const {
+  auto ib = md->GetBoundsI(IndexDomain::interior);
+  auto jb = md->GetBoundsJ(IndexDomain::interior);
+  auto kb = md->GetBoundsK(IndexDomain::interior);
+  auto bnds = AMRBounds(ib, jb, kb);
+  auto dims = md->GetMeshPointer()->resolved_packages->FieldMetadata(field).Shape();
+  int n5(0), n4(0);
+  if (dims.size() > 2) {
+    n5 = dims[1];
+    n4 = dims[2];
+  } else if (dims.size() > 1) {
+    n5 = dims[0];
+    n4 = dims[1];
+  }
+  const int idx = comp4 + n4 * (comp5 + n5 * comp6);
+  Refinement::SecondDerivative(bnds, md, field, idx, amr_tags, refine_criteria,
+                               derefine_criteria, max_level);
 }
 
 } // namespace parthenon