+++ src/amr_criteria/refinement_package.cpp
@@ -20,19 +20,25 @@
 #include <utility>
 
 #include "amr_criteria/amr_criteria.hpp"
+#include "interface/make_pack_descriptor.hpp"
 #include "interface/mesh_data.hpp"
 #include "interface/meshblock_data.hpp"
 #include "interface/state_descriptor.hpp"
+#include "kokkos_abstraction.hpp"
 #include "mesh/mesh.hpp"
 #include "mesh/mesh_refinement.hpp"
 #include "mesh/meshblock.hpp"
 #include "parameter_input.hpp"
+#include "utils/instrument.hpp"
 
 namespace parthenon {
 namespace Refinement {
 
 std::shared_ptr<StateDescriptor> Initialize(ParameterInput *pin) {
   auto ref = std::make_shared<StateDescriptor>("Refinement");
+  bool check_refine_mesh =
+      pin->GetOrAddBoolean("parthenon/mesh", "CheckRefineMesh", true);
+  ref->AddParam("check_refine_mesh", check_refine_mesh);
 
   int numcrit = 0;
   while (true) {
@@ -48,7 +54,30 @@ std::shared_ptr<StateDescriptor> Initialize(ParameterInput *pin) {
   return ref;
 }
 
-AmrTag CheckAllRefinement(MeshBlockData<Real> *rc) {
+ParArray1D<AmrTag> CheckAllRefinement(MeshData<Real> *md) {
+  const int nblocks = md->NumBlocks();
+  Mesh *pm = md->GetMeshPointer();
+  auto amr_tags = pm->GetAmrTags();
+  Kokkos::deep_copy(amr_tags.KokkosView(), AmrTag::derefine);
+
+  static const bool check_refine_mesh =
+      pm->packages.Get("Refinement")->Param<bool>("check_refine_mesh");
+
+  for (auto &pkg : pm->packages.AllPackages()) {
+    auto &desc = pkg.second;
+    desc->CheckRefinement(md, amr_tags);
+
+    if (check_refine_mesh) {
+      for (auto &amr : desc->amr_criteria) {
+        (*amr)(md, amr_tags);
+      }
+    }
+  }
+
+  return amr_tags;
+}
+
+AmrTag CheckAllRefinement(MeshBlockData<Real> *rc, const AmrTag &level) {
   // Check all refinement criteria and return the maximum recommended change in
   // refinement level:
   //   delta_level = -1 => recommend derefinement
@@ -62,8 +91,11 @@ AmrTag CheckAllRefinement(MeshBlockData<Real> *rc) {
   //       neighboring blocks.  Similarly for "do nothing"
   PARTHENON_INSTRUMENT
   MeshBlock *pmb = rc->GetBlockPointer();
-  // delta_level holds the max over all criteria.  default to derefining.
-  AmrTag delta_level = AmrTag::derefine;
+  static const bool check_refine_mesh =
+      pmb->packages.Get("Refinement")->Param<bool>("check_refine_mesh");
+  // delta_level holds the max over all criteria.  default to derefining, or level from
+  // MeshData check.
+  AmrTag delta_level = level;
   for (auto &pkg : pmb->packages.AllPackages()) {
     auto &desc = pkg.second;
     delta_level = std::max(delta_level, desc->CheckRefinement(rc));
@@ -71,6 +103,7 @@ AmrTag CheckAllRefinement(MeshBlockData<Real> *rc) {
       // since 1 is the max, we can return without having to look at anything else
       return AmrTag::refine;
     }
+    if (check_refine_mesh) continue;
     // call parthenon criteria that were registered
     for (auto &amr : desc->amr_criteria) {
       // get the recommended change in refinement level from this criteria
@@ -150,9 +183,111 @@ AmrTag SecondDerivative(const AMRBounds &bnds, const ParArray3D<Real> &q,
   return AmrTag::same;
 }
 
-void SetRefinement_(MeshBlockData<Real> *rc) {
+void FirstDerivative(const AMRBounds &bnds, MeshData<Real> *md, const std::string &field,
+                     const int &idx, ParArray1D<AmrTag> &amr_tags,
+                     const Real refine_criteria_, const Real derefine_criteria_,
+                     const int max_level_) {
+  const auto desc =
+      MakePackDescriptor(md->GetMeshPointer()->resolved_packages.get(), {field});
+  auto pack = desc.GetPack(md);
+  const int ndim = md->GetMeshPointer()->ndim;
+  const int nvars = pack.GetMaxNumberOfVars();
+
+  const Real refine_criteria = refine_criteria_;
+  const Real derefine_criteria = derefine_criteria_;
+  const int max_level = max_level_;
+  const int var = idx;
+  auto scatter_tags = amr_tags.ToScatterView<Kokkos::Experimental::ScatterMax>();
+  par_for_outer(
+      PARTHENON_AUTO_LABEL, 0, 0, 0, pack.GetNBlocks() - 1, bnds.ks, bnds.ke, bnds.js,
+      bnds.je,
+      KOKKOS_LAMBDA(team_mbr_t team_member, const int b, const int k, const int j) {
+        Real maxd = 0.;
+        par_reduce_inner(
+            inner_loop_pattern_ttr_tag, team_member, bnds.is, bnds.ie,
+            [&](const int i, Real &maxder) {
+              Real scale = std::abs(pack(b, var, k, j, i));
+              Real d = 0.5 *
+                       std::abs((pack(b, var, k, j, i + 1) - pack(b, var, k, j, i - 1))) /
+                       (scale + TINY_NUMBER);
+              maxder = (d > maxder ? d : maxder);
+              if (ndim > 1) {
+                d = 0.5 *
+                    std::abs((pack(b, var, k, j + 1, i) - pack(b, var, k, j - 1, i))) /
+                    (scale + TINY_NUMBER);
+                maxder = (d > maxder ? d : maxder);
+              }
+              if (ndim > 2) {
+                d = 0.5 *
+                    std::abs((pack(b, var, k + 1, j, i) - pack(b, var, k - 1, j, i))) /
+                    (scale + TINY_NUMBER);
+                maxder = (d > maxder ? d : maxder);
+              }
+            },
+            Kokkos::Max<Real>(maxd));
+        auto tags_access = scatter_tags.access();
+        auto flag = AmrTag::same;
+        if (maxd > refine_criteria && pack.GetLevel(b, 0, 0, 0) < max_level)
+          flag = AmrTag::refine;
+        if (maxd < derefine_criteria) flag = AmrTag::derefine;
+        tags_access(b).update(flag);
+      });
+  amr_tags.ContributeScatter(scatter_tags);
+}
+
+void SecondDerivative(const AMRBounds &bnds, MeshData<Real> *md, const std::string &field,
+                      const int &idx, ParArray1D<AmrTag> &amr_tags,
+                      const Real refine_criteria_, const Real derefine_criteria_,
+                      const int max_level_) {
+  const auto desc =
+      MakePackDescriptor(md->GetMeshPointer()->resolved_packages.get(), {field});
+  auto pack = desc.GetPack(md);
+  const int ndim = md->GetMeshPointer()->ndim;
+  const int nvars = pack.GetMaxNumberOfVars();
+
+  const Real refine_criteria = refine_criteria_;
+  const Real derefine_criteria = derefine_criteria_;
+  const int max_level = max_level_;
+  const int var = idx;
+  auto scatter_tags = amr_tags.ToScatterView<Kokkos::Experimental::ScatterMax>();
+  par_for_outer(
+      PARTHENON_AUTO_LABEL, 0, 0, 0, pack.GetNBlocks() - 1, bnds.ks, bnds.ke, bnds.js,
+      bnds.je,
+      KOKKOS_LAMBDA(team_mbr_t team_member, const int b, const int k, const int j) {
+        Real maxd = 0.;
+        par_reduce_inner(
+            inner_loop_pattern_ttr_tag, team_member, bnds.is, bnds.ie,
+            [&](const int i, Real &maxder) {
+              Real aqt = std::abs(pack(b, var, k, j, i)) + TINY_NUMBER;
+              Real qavg = 0.5 * (pack(b, var, k, j, i + 1) + pack(b, var, k, j, i - 1));
+              Real d = std::abs(qavg - pack(b, var, k, j, i)) / (std::abs(qavg) + aqt);
+              maxder = (d > maxder ? d : maxder);
+              if (ndim > 1) {
+                qavg = 0.5 * (pack(b, var, k, j + 1, i) + pack(b, var, k, j - 1, i));
+                d = std::abs(qavg - pack(b, var, k, j, i)) / (std::abs(qavg) + aqt);
+                maxder = (d > maxder ? d : maxder);
+              }
+              if (ndim > 2) {
+                qavg = 0.5 * (pack(b, var, k + 1, j, i) + pack(b, var, k - 1, j, i));
+                d = std::abs(qavg - pack(b, var, k, j, i)) / (std::abs(qavg) + aqt);
+                maxder = (d > maxder ? d : maxder);
+              }
+            },
+            Kokkos::Max<Real>(maxd));
+        auto tags_access = scatter_tags.access();
+        auto flag = AmrTag::same;
+        if (maxd > refine_criteria && pack.GetLevel(b, 0, 0, 0) < max_level)
+          flag = AmrTag::refine;
+        if (maxd < derefine_criteria) flag = AmrTag::derefine;
+        tags_access(b).update(flag);
+      });
+  amr_tags.ContributeScatter(scatter_tags);
+}
+
+void SetRefinement_(MeshBlockData<Real> *rc,
+                    const AmrTag &delta_level = AmrTag::derefine) {
   auto pmb = rc->GetBlockPointer();
-  pmb->pmr->SetRefinement(CheckAllRefinement(rc));
+  pmb->pmr->SetRefinement(CheckAllRefinement(rc, delta_level));
 }
 
 template <>
@@ -163,10 +298,13 @@ TaskStatus Tag(MeshBlockData<Real> *rc) {
 }
 
 template <>
-TaskStatus Tag(MeshData<Real> *rc) {
+TaskStatus Tag(MeshData<Real> *md) {
   PARTHENON_INSTRUMENT
-  for (int i = 0; i < rc->NumBlocks(); i++) {
-    SetRefinement_(rc->GetBlockData(i).get());
+  ParArray1D<AmrTag> amr_tags = CheckAllRefinement(md);
+  auto amr_tags_h = amr_tags.GetHostMirrorAndCopy();
+
+  for (int i = 0; i < md->NumBlocks(); i++) {
+    SetRefinement_(md->GetBlockData(i).get(), amr_tags_h(i));
   }
   return TaskStatus::complete;
 }