+++ src/amr_criteria/refinement_package.hpp
@@ -37,13 +37,24 @@ std::shared_ptr<StateDescriptor> Initialize(ParameterInput *pin);
 template <typename T>
 TaskStatus Tag(T *rc);
 
-AmrTag CheckAllRefinement(MeshBlockData<Real> *rc);
+AmrTag CheckAllRefinement(MeshBlockData<Real> *rc, const AmrTag &level);
+ParArray1D<AmrTag> CheckAllRefinement(MeshData<Real> *md);
 
 AmrTag FirstDerivative(const AMRBounds &bnds, const ParArray3D<Real> &q,
                        const Real refine_criteria, const Real derefine_criteria);
 
+void FirstDerivative(const AMRBounds &bnds, MeshData<Real> *md, const std::string &field,
+                     const int &idx, ParArray1D<AmrTag> &amr_tags,
+                     const Real refine_criteria_, const Real derefine_criteria_,
+                     const int max_level_);
+
 AmrTag SecondDerivative(const AMRBounds &bnds, const ParArray3D<Real> &q,
                         const Real refine_criteria, const Real derefine_criteria);
+
+void SecondDerivative(const AMRBounds &bnds, MeshData<Real> *md, const std::string &field,
+                      const int &idx, ParArray1D<AmrTag> &amr_tags,
+                      const Real refine_criteria_, const Real derefine_criteria_,
+                      const int max_level_);
 
 } // namespace Refinement